+++ Common/Status.h
@@ -3,10 +3,13 @@
 
 #include "Common/Common.h"
 
+namespace DLS {
+
 enum class Status : uint8_t {
 	Failure = 0,
 	Success = 1
 };
 
+} // namespace DLS
 
 #endif /* __COMMON_STATUS_H__ */
