--- Configuration.h
@@ -1,15 +1,0 @@
-#ifndef __CONFIGURATION_H__
-#define __CONFIGURATION_H__
-
-#include "Common/Common.h"
-
-struct Configuration {
-
-	// room temperature for now
-	static constexpr double kKelvinTemperature = 298;
-
-
-};
-
-
-#endif /* __CONFIGURATION_H__ */
