--- Constant.h
@@ -1,15 +1,0 @@
-#ifndef __CONSTANT_H__
-#define __CONSTANT_H__
-
-#include "Common/Common.h"
-
-struct Constant {
-
-	// in m^2 * kg / ( s^2 * K )
-	static constexpr double kBoltzmannConstant = 1.38064852e-23;
-
-
-};
-
-
-#endif /* __CONSTANT_H__ */
