+++ Math/Point.h
@@ -3,6 +3,9 @@
 
 #include "Common/Common.h"
 
+namespace DLS {
+namespace Math {
+
 template< typename T >
 struct Point2D {
     Point2D( T _x, T _y )
@@ -30,5 +33,8 @@ struct Point3D {
     T y;
     T z;
 };
+
+} // namespace Math
+} // namespace DLS
 
 #endif /* __MATH_POINT_H__ */
