--- ParticleType.h
@@ -1,11 +1,0 @@
-#ifndef __PARTICLE_TYPE_H__
-#define __PARTICLE_TYPE_H__
-
-#include "Common/Common.h"
-
-enum class ParticleType {
-    PM2_5,
-    PM10
-};
-
-#endif /* __PARTICLE_TYPE_H__ */
