+++ Physics/ParticleType.h
@@ -1,0 +1,17 @@
+#ifndef __PHYSICS_PARTICLE_TYPE_H__
+#define __PHYSICS_PARTICLE_TYPE_H__
+
+#include "Common/Common.h"
+
+namespace DLS {
+namespace Physics {
+
+enum class ParticleType {
+    PM2_5,
+    PM10
+};
+
+} // namespace Physics
+} // namespace DLS
+
+#endif /* __PHYSICS_PARTICLE_TYPE_H__ */
