+++ Utils/StringToNumber.h
@@ -3,11 +3,17 @@
 
 #include "Common/Common.h"
 
+namespace DLS {
+namespace Utils {
+
 template< typename IntegralType >
 IntegralType StringToIntegralType( const std::string & input_string );
 
 template< typename FloatingType >
 FloatingType StringToFloatingType( const std::string & input_string );
+
+} // namespace Utils
+} // namespace DLS
 
 #include "Utils/StringToNumber.inl"
