+++ Utils/StringToNumber.inl
@@ -1,6 +1,9 @@
 #ifndef __UTILS_STRING_TO_NUMBER_INL__
 #define __UTILS_STRING_TO_NUMBER_INL__
 
+namespace DLS {
+namespace Utils {
+
 // specifically for integral types
 template< typename IntegralType >
 IntegralType StringToIntegralType( const std::string & input_string )
@@ -14,5 +17,8 @@ FloatingType StringToFloatingType( const std::string & input_string )
 {
 	return static_cast< FloatingType >( std::stod( input_string ) );	
 }
+
+} // namespace Utils
+} // namespace DLS
 
 #endif /* __UTILS_STRING_TO_NUMBER_INL__ */
